+++ tests/testthat/test_read_tracts.R
@@ -20,7 +20,7 @@ test_that("read_tracts", {
   test2 <- read_tracts(year = 2010, dataset = 'Basico', as_data_frame = TRUE)
   testthat::expect_true(is(test2, "data.frame"))
 
-  # different data sets
+  # 2010 different data sets
   ## check if file have been downloaded
   tbls <- c('Basico', 'Domicilio', 'DomicilioRenda', 'Entorno',
              'ResponsavelRenda', 'Responsavel', 'PessoaRenda', 'Pessoa')
@@ -28,6 +28,14 @@ test_that("read_tracts", {
   lapply(X=tbls, FUN = function(y){ # y = 'Pessoa'     y = 'Basico'  y = 'Entorno'
     tmp_d <- read_tracts(year = 2010, dataset = y)
     testthat::expect_true( nrow(tmp_d) >= 303000)
+  } )
+
+  # 2022 different data sets
+  ## check if file have been downloaded
+  tbls <- c('Preliminares')
+  lapply(X=tbls, FUN = function(y){ # y = 'Preliminares'
+    tmp_d <- read_tracts(year = 2022, dataset = y)
+    testthat::expect_true( nrow(tmp_d) == 452340)
   } )
 
 
@@ -54,7 +62,7 @@ test_that("read_tracts", {
   testthat::expect_error(read_tracts(year=999, dataset='Basico'))
   testthat::expect_error(read_tracts(year=999, dataset='Basico'))
   testthat::expect_error(read_tracts(year=2010, dataset='banana'))
-  testthat::expect_error(read_tracts(year=2010, dataset='banana'))
+  testthat::expect_error(read_tracts(year=2022, dataset='banana'))
 
   testthat::expect_error(read_tracts(year=2010, dataset='Basico', showProgress = 'banana' ))
   testthat::expect_error(read_tracts(year=2010, dataset='Basico', cache = 'banana' ))